requirements.txt
# 大模型调用
openai>=1.0.0

# 数据处理
pandas>=2.0.0

# 职业图谱
networkx>=3.0

# 报告导出
python-docx>=1.0.0
reportlab>=4.0.0

# 简历解析（PDF/Word 文本提取）
pdfplumber>=0.10.0
python-dotenv>=1.0.0

# Web 界面
streamlit>=1.28.0
plotly>=5.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# 工具
requests>=2.28.0
